Add working auth flow (still have to fix progressbar though)

+++ app/src/main/java/org/tetawex/vkphotoviewer/app/model/interactor/FriendsDetailInteractor.kt
@@ -7,5 +7,5 @@ import org.tetawex.vkphotoviewer.app.model.repository.api.dto.FriendsDetail
  * Created by tetawex on 20.07.2018.
  */
 interface FriendsDetailInteractor {
-    fun getFriendsDetail(id: Int): Single<FriendsDetail>
+    fun getFriendsDetail(id: Long): Single<FriendsDetail>
 }
